citifix_backend/requirements.txt
asgiref==3.10.0
black==25.9.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
PyJWT==2.10.1
python-decouple==3.8
pytokens==0.2.0
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
